De-clutter consumer_group.rs by breaking out builder
